requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
yfinance>=0.2.40
xgboost>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
ta>=0.11.0
joblib>=1.3.0
statsmodels>=0.14.0
streamlit>=1.35.0
Typer>=0.12.0
rich>=13.7.0
pyarrow>=15.0.0
python-docx>=1.1.2
google-api-python-client>=2.136.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
reportlab>=4.2.5
alpaca-py>=0.26.0
python-dotenv>=1.0.1
websockets>=13.0.0
kiteconnect>=5.0.0
